feat(build): added turbopack in dev & fixed circular dependency
