+++ sim/blocks/blocks/agent.ts
@@ -2,7 +2,7 @@ import { AgentIcon } from '@/components/icons'
 import { createLogger } from '@/lib/logs/console-logger'
 import { isHosted } from '@/lib/environment'
 import { useOllamaStore } from '@/stores/ollama/store'
-import { getAllBlocks } from '@/blocks'
+import { getAllBlocks } from '@/blocks/registry'
 import { MODELS_TEMP_RANGE_0_1, MODELS_TEMP_RANGE_0_2 } from '@/providers/model-capabilities'
 import { getAllModelProviders, getBaseModelProviders } from '@/providers/utils'
 import { ToolResponse } from '@/tools/types'
