+++ sim/sentry.server.config.ts
@@ -4,19 +4,19 @@
 
 import * as Sentry from "@sentry/nextjs"
 
-Sentry.init({
-  dsn: process.env.NEXT_PUBLIC_SENTRY_DSN || undefined,
-  enabled: process.env.NODE_ENV === 'production',
-  environment: process.env.NODE_ENV || 'development',
-  tracesSampleRate: process.env.NODE_ENV === 'production' ? 0.2 : 1,
-  debug: process.env.NODE_ENV === 'development',
-  
-  beforeSend(event) {
-    if (process.env.NODE_ENV !== 'production') return null
+// Completely skip Sentry initialization in development
+if (process.env.NODE_ENV === 'production') {
+  Sentry.init({
+    dsn: process.env.NEXT_PUBLIC_SENTRY_DSN || undefined,
+    enabled: true,
+    environment: process.env.NODE_ENV || 'development',
+    tracesSampleRate: 0.2,
     
-    if (event.request && typeof event.request === 'object') {
-      (event.request as any).ip = null
-    }
-    return event
-  },
-})
+    beforeSend(event) {
+      if (event.request && typeof event.request === 'object') {
+        (event.request as any).ip = null
+      }
+      return event
+    },
+  })
+}
